Initial commit
This script will retrieve the Attribute Sync config for a specified source, and iterate through all correlated accounts for the source, grabbing the source account attribute value and the corresponding Identity Attribute value, then comparing the two to see if they match. The results can then either be output to a CSV file or to the clipboard.